+++ src/test/scala/dsptools/numbers/AbsSpec.scala
@@ -29,16 +29,16 @@ class AbsSpec extends FreeSpec with Matchers {
 }
 
 class DoesAbsTester(c: DoesAbs[UInt, SInt, FixedPoint]) extends DspTester(c) {
-  for(i <- 0.0 to 15.0 by 1.0) {
-    poke(c.io.uIn, i)
-    expect(c.io.uAbsGrow, i)
-    expect(c.io.uAbsWrap, i)
+  for(i <- BigDecimal(0.0) to 15.0 by 1.0) {
+    poke(c.io.uIn, i.toDouble)
+    expect(c.io.uAbsGrow, i.toDouble)
+    expect(c.io.uAbsWrap, i.toDouble)
     step(1)
   }
-  for(i <- -7.0 to 7.0 by 1.0) {
-    poke(c.io.sIn, i)
-    expect(c.io.sAbsGrow, i.abs)
-    expect(c.io.sAbsWrap, i.abs)
+  for(i <- BigDecimal(-7.0) to 7.0 by 1.0) {
+    poke(c.io.sIn, i.toDouble)
+    expect(c.io.sAbsGrow, i.abs.toDouble)
+    expect(c.io.sAbsWrap, i.abs.toDouble)
     step(1)
   }
   poke(c.io.sIn, -8.0)
@@ -47,10 +47,10 @@ class DoesAbsTester(c: DoesAbs[UInt, SInt, FixedPoint]) extends DspTester(c) {
 
   val increment = 0.25
 
-  for(i <- -3.75 to 3.75 by increment) {
-    poke(c.io.fIn, i)
-    expect(c.io.fAbsGrow, i.abs)
-    expect(c.io.fAbsWrap, i.abs)
+  for(i <- BigDecimal(-3.75) to 3.75 by increment) {
+    poke(c.io.fIn, i.toDouble)
+    expect(c.io.fAbsGrow, i.abs.toDouble)
+    expect(c.io.fAbsWrap, i.abs.toDouble)
     step(1)
   }
   poke(c.io.fIn, -4.0)
